requirements-dev.txt
black
coverage
parameterized
pylint
pytest
pytest-cov
pytest-html
pytest_redis
vcrpy
vulture